ProductController: Done - get all products

+++ src/main/java/com/example/hengapi/model/Products.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class Products {
     private Integer productId;
     private String productName;
-    private int productPrice;
+    private float productPrice;
 }
